% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sticker.R
\name{geom_hexagon}
\alias{geom_hexagon}
\title{geom_hexagon}
\usage{
geom_hexagon(size = 1.2, fill = "#1881C2", color = "#87B13F")
}
\arguments{
\item{size}{size of border}

\item{fill}{color of hexagon}

\item{color}{color of border}
}
\value{
hexagon layer
}
\description{
geom layer of hexagon
}
\author{
guangchuang yu
}

