% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sticker.R
\name{geom_pkgname}
\alias{geom_pkgname}
\title{geom_pkgname}
\usage{
geom_pkgname(package, x = 1, y = 1.4, color = "#FFFFFF",
  family = "Aller_Rg", size = 8, ...)
}
\arguments{
\item{package}{package name}

\item{x}{x position}

\item{y}{y position}

\item{color}{color}

\item{family}{font family}

\item{size}{font size}

\item{...}{addition parameters passed to geom_text()}
}
\value{
package name layer
}
\description{
add package name to sticker
}
\author{
Guangchuang Yu
}
