% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sticker.R
\name{geom_url}
\alias{geom_url}
\title{geom_url}
\usage{
geom_url(
  url = "www.bioconductor.org",
  x = 1,
  y = 0.08,
  family = "Aller_Rg",
  size = 1.5,
  color = "black",
  angle = 30,
  hjust = 0,
  ...
)
}
\arguments{
\item{url}{url}

\item{x}{x position of url}

\item{y}{y position of url}

\item{family}{font family}

\item{size}{size of url}

\item{color}{color of url}

\item{angle}{angle of url, default is 30}

\item{hjust}{horizontal adjustment}

\item{...}{additional parameters to geom_text}
}
\value{
geom layer
}
\description{
add url at the lower border of the sticker
}
\author{
Guangchuang Yu
}
