% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sticker.R
\name{hexagon}
\alias{hexagon}
\title{hexagon}
\usage{
hexagon(size = 1.2, fill = "#1881C2", color = "#87B13F")
}
\arguments{
\item{size}{size of border}

\item{fill}{color of hexagon}

\item{color}{border color of hexagon}
}
\value{
hexagon
}
\description{
empty hexagon
}
\author{
guangchuang yu
}
