% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexSticker-package.R
\docType{package}
\name{hexSticker-package}
\alias{hexSticker}
\alias{hexSticker-package}
\title{hexSticker: Create Hexagon Sticker in R}
\description{
Helper functions for creating reproducible hexagon sticker purely in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GuangchuangYu/hexSticker}
  \item Report bugs at \url{https://github.com/GuangchuangYu/hexSticker/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com}

Other contributors:
\itemize{
  \item Laurent Gatto \email{lg390@cam.ac.uk} [contributor]
  \item Johannes Rainer \email{johannes.rainer@eurac.edu} [contributor]
  \item Sebastian Gibb \email{mail@sebastiangibb.de} [contributor]
  \item Robert Corty \email{rcorty@gmail.com} [contributor]
  \item Thomas Neitmann \email{th.neitmann@gmail.com} [contributor]
}

}
\keyword{internal}
