% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sticker.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{theme_transparent}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggtree}{\code{\link[ggtree]{theme_transparent}}}
}}
\keyword{internal}

