% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sticker.R
\name{save_sticker}
\alias{save_sticker}
\title{save_sticker}
\usage{
save_sticker(filename, sticker = last_plot())
}
\arguments{
\item{filename}{file name}

\item{sticker}{sticker}
}
\description{
save sticker to file
}
\author{
guangchuang yu
}
