% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sticker.R
\name{theme_sticker}
\alias{theme_sticker}
\title{theme_sticker}
\usage{
theme_sticker(...)
}
\arguments{
\item{...}{additional parameters passed to theme()}
}
\value{
theme for sticker
}
\description{
sticker theme
}
\author{
guangchuang yu
}

